fixed assert dialog in VC++ CMake multiwindow included to Demo significantly improved EventDispatcher performance added error handler callback to oxygine::log added std::function ThreadMessages callback updated to latest Marmalade and fixed build minor fixes

+++ oxygine/src/core/s3e/MyHttp.cpp
@@ -258,7 +258,7 @@ void MyHttp::gotData(int size)
 int MyHttp::_gotData (void* systemData, void* userData)
 {
 	MyHttp *http = (MyHttp *)userData;
-	http->gotData((int)systemData);
+	http->gotData((size_t)systemData);
 	return 0;
 }
